Enhanced: - Added precompiled headers support for gcc - Fixed a gcc warning
git-svn-id: http://opentyrian.googlecode.com/svn@812 194990a9-7324-0410-9a02-4386df949664

+++ enhanced/trunk/src/scroller.cpp
@@ -160,7 +160,7 @@ void scroller_sine( const about_text_type text[] )
 			}
 		}
 
-		for (int i = 0; i < visible_lines; i++)
+		for (int i = 0; unsigned(i) < visible_lines; i++)
 		{
 			if (current_line + i >= 0)
 			{
